improve perf of integer stats with no-op logsquare
turns out nearly all the time taken in adding values is
because of the log involved in computing the log sum of squares.
Since this is not a stat that I think is commonly needed, taking
this out as the main purpose of this class is to make common stats on integers
really fast.